SpotLightをファイルに切り出し
