+++ app/src/main/java/com/yuta/cocktailmaster/ui/screen/TopScreen.kt
@@ -212,97 +212,6 @@ fun TopScreen(
     }
 }
 
-@OptIn(ExperimentalTextApi::class, ExperimentalUnitApi::class)
-@Composable
-fun SpotLight(
-    rect: Rect,
-    topAppBarSize: IntSize,
-    text: String = "",
-    textAreaPosition: TextAreaPosition,
-    isLast: Boolean,
-    onAreaTapped: () -> Unit,
-) {
-    println("spotLight isLast: $isLast")
-    val actual = rect.translate(
-        translateX = 0f,
-        translateY = -topAppBarSize.height.toFloat()
-    ).inflate(8f)
-    val textMeasure = rememberTextMeasurer()
-
-    Canvas(
-        modifier = Modifier
-            .fillMaxSize()
-            .pointerInput(Unit) {
-                detectTapGestures {
-                    onAreaTapped()
-                }
-            }
-    ) {
-        val spotlightPath = Path().apply {
-            addRect(actual)
-        }
-        clipPath(
-            path = spotlightPath,
-            clipOp = ClipOp.Difference
-        ) {
-            drawRect(color = Color.Black.copy(alpha = 0.8f))
-        }
-        if(isLast) {
-            val width = size.width
-            val height = size.height
-            val textLayoutResult = textMeasure.measure(
-                text = AnnotatedString("さぁ使ってみましょう!"),
-                style = TextStyle(
-                    fontSize = 26.sp,
-                    color = Color.White,
-                    lineHeight = TextUnit(30f, TextUnitType.Sp)
-                )
-            )
-            val textSize = textLayoutResult.size
-            println("width: $width height: $height")
-            println("textWidth: ${textSize.width} textHeight: ${textSize.height}")
-            drawText(
-                textMeasure,
-                text = "さぁ使ってみましょう!",
-                topLeft = Offset(
-                    x = (width - textSize.width) / 2,
-                    y = (height - textSize.height) / 2
-                ),
-                style = TextStyle(
-                    fontSize = 26.sp,
-                    color = Color.White,
-                    lineHeight = TextUnit(30f, TextUnitType.Sp)
-                )
-            )
-        } else {
-
-            val yPos = when (textAreaPosition) {
-                TextAreaPosition.ABOVE -> actual.top - 100
-                TextAreaPosition.BELOW -> actual.bottom + 30
-            }
-            drawText(
-                textMeasure,
-                text = text,
-                topLeft = Offset(
-                    x = 40f,
-                    y = yPos
-                ),
-                style = TextStyle(
-                    fontSize = 26.sp,
-                    color = Color.White,
-                    lineHeight = TextUnit(30f, TextUnitType.Sp)
-                )
-            )
-        }
-
-    }
-}
-
-enum class TextAreaPosition {
-    ABOVE,
-    BELOW,
-}
-
 @Composable
 fun DeleteConfirmDialog(
     title: String,
